requirements.txt
Django==5.2.6
django-allauth==0.63.6
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
Pillow==10.4.0
python-decouple==3.8
whitenoise==6.8.1
gunicorn==23.0.0
PyJWT==2.8.0
psycopg2-binary==2.9.9